backend/requirements.txt
fastapi==0.115.2
uvicorn==0.34.2
pydantic==2.11.5
python-dotenv==1.1.0
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
